Rename database module to be more specific: sqlite
